optimiseur_rl/requirements.txt
gym==0.26.2
gymnasium==0.29.1
Markdown==3.6
matplotlib==3.8.3
numpy==1.26.4
PyYAML==6.0.1
python-dateutil==2.8.2
Sphinx==7.2.6
stable_baselines3==2.3.0a2
torch==2.2.1
tqdm==4.66.2

